Creation of Box DTOs

+++ src/main/java/fr/tours/etu/boiteletre/Service/BoxService.java
@@ -1,48 +1,72 @@
 package fr.tours.etu.boiteletre.Service;
 
+import fr.tours.etu.boiteletre.DTO.DtoForBox.BoxDTO;
+import fr.tours.etu.boiteletre.DTO.DtoForBox.ResponseBoxDTO;
+import fr.tours.etu.boiteletre.DTO.ReservationDTO;
+import fr.tours.etu.boiteletre.MappStruct.BoxMapper;
 import fr.tours.etu.boiteletre.Model.Box;
 import fr.tours.etu.boiteletre.Repository.BoxRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class BoxService {
 
     private final BoxRepository boxRepository;
+    private final BoxMapper boxMapper;
 
-    @Autowired
-    public BoxService(BoxRepository boxRepository){
-        this.boxRepository = boxRepository;
+    public ResponseBoxDTO createBox(BoxDTO boxDTO){
+
+        Box box = boxMapper.dtoToBox(boxDTO);
+        Box saveBox = boxRepository.save(box);
+
+        BoxDTO dtoBox = boxMapper.boxToDto(saveBox);
+
+        return new ResponseBoxDTO(dtoBox);
     }
 
 
-    public Box createBox(Box box){
-        return boxRepository.save(box);
+    public List<ResponseBoxDTO> getAllBox(){
+        List<Box> boxList = boxRepository.findAll();
+        List<ResponseBoxDTO> responseBoxDTOList = new ArrayList<>();
+
+        if (!boxList.isEmpty()){
+            for (Box b : boxList) {
+                responseBoxDTOList.add(new ResponseBoxDTO(boxMapper.boxToDto(b)));
+            }
+        }
+        return responseBoxDTOList;
     }
 
+    public ResponseBoxDTO getBoxById(int id){
+        Box box = boxRepository.findById(id).orElseThrow(()-> new IllegalArgumentException("The box with the id : " + id + " doesn't exist!"));
+
+        return new ResponseBoxDTO(boxMapper.boxToDto(box));
+    }
+
+    public ResponseBoxDTO updateBox(int id, BoxDTO boxDTO){
+        Box box = boxRepository.findById(id).orElseThrow(()-> new IllegalArgumentException("The box with the id : " + id + " doesn't exist!"));
+
+        box.setName(boxDTO.getName());
+        box.setQuantity(boxDTO.getQuantity());
+        box.setDescription(boxDTO.getDescription());
+
+        return new ResponseBoxDTO(boxMapper.boxToDto(boxRepository.save(box)));
+    }
 
     public void deleteBoxById(int id){
-        boxRepository.deleteById(id);
-    }
 
-    public List<Box> getAllBox(){
-        return boxRepository.findAll();
-    }
+        if (boxRepository.existsById(id)){
+            boxRepository.deleteById(id);
+        }else{
+            throw new IllegalArgumentException("The box with the id : " + id + " doesn't exist!");
+        }
 
-    public Box getBoxById(int id){
-        return boxRepository.findById(id).orElse(null);
-    }
-
-    public Box updateBox(int id, String name, int quantity, String description){
-        Box box = boxRepository.findById(id).orElseThrow(()-> new IllegalArgumentException("Boite not found !"));
-
-        box.setName(name);
-        box.setQuantity(quantity);
-        box.setDescription(description);
-
-        return boxRepository.save(box);
     }
 
 }
